+++ tokens/tests/test_views.py
@@ -1,11 +1,13 @@
-"""View tests for the tokens app"""
+"""UI view tests for the tokens app"""
 
-
+from django.contrib.messages import get_messages
 from django.urls import reverse
 
 from knox.models import AuthToken
 
 from test_plus.test import TestCase
+
+from tokens.views import TOKEN_CREATE_MSG, TOKEN_DELETE_MSG
 
 
 class TestUserTokenListView(TestCase):
@@ -18,18 +20,18 @@ class TestUserTokenListView(TestCase):
         self.tokens = [AuthToken.objects.create(self.user, None)]
 
     def test_list_empty(self):
-        """Test that rendering the list view without any tokens works"""
+        """Test rendering the list view without tokens"""
         with self.login(self.user):
             response = self.get('tokens:list')
-        self.response_200(response)
+        self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.context['object_list']), 0)
 
     def test_list_one(self):
-        """Test that rendering the list view with one token works"""
+        """Test rendering the list view with one token"""
         self._make_token()
         with self.login(self.user):
             response = self.get('tokens:list')
-        self.response_200(response)
+        self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.context['object_list']), 1)
 
 
@@ -40,27 +42,35 @@ class TestUserTokenCreateView(TestCase):
         self.user = self.make_user()
 
     def test_get(self):
-        """Test that showing the creation form works"""
+        """Test rendering the creation form"""
         with self.login(self.user):
             response = self.get('tokens:create')
-        self.response_200(response)
+        self.assertEqual(response.status_code, 200)
         self.assertIsNotNone(response.context["form"])
 
-    def test_post_success_no_ttl(self):
-        """Test creating an authentication token with TTL=0 works"""
+    def test_post_no_ttl(self):
+        """Test creating an authentication token with TTL=0"""
         self.assertEqual(AuthToken.objects.count(), 0)
         with self.login(self.user):
             response = self.post('tokens:create', data={'ttl': 0})
-        self.response_200(response)
+        self.assertEqual(response.status_code, 200)
         self.assertEqual(AuthToken.objects.count(), 1)
+        self.assertEqual(
+            list(get_messages(response.wsgi_request))[0].message,
+            TOKEN_CREATE_MSG,
+        )
 
-    def test_post_success_with_ttl(self):
-        """Test creating an authentication token with TTL != 0 works"""
+    def test_post_ttl(self):
+        """Test creating an authentication token with TTL != 0"""
         self.assertEqual(AuthToken.objects.count(), 0)
         with self.login(self.user):
             response = self.post('tokens:create', data={'ttl': 10})
-        self.response_200(response)
+        self.assertEqual(response.status_code, 200)
         self.assertEqual(AuthToken.objects.count(), 1)
+        self.assertEqual(
+            list(get_messages(response.wsgi_request))[0].message,
+            TOKEN_CREATE_MSG,
+        )
 
 
 class TestUserTokenDeleteView(TestCase):
@@ -72,17 +82,21 @@ class TestUserTokenDeleteView(TestCase):
         self.token = AuthToken.objects.first()
 
     def test_get(self):
-        """Test that showing the deletion form works"""
+        """Test rendering the deletion form"""
         with self.login(self.user):
             response = self.get('tokens:delete', pk=self.token.pk)
-        self.response_200(response)
+        self.assertEqual(response.status_code, 200)
 
-    def test_post_success(self):
-        """Test that deleting a token works"""
+    def test_post(self):
+        """Test token deletion"""
         self.assertEqual(AuthToken.objects.count(), 1)
         with self.login(self.user):
             response = self.post('tokens:delete', pk=self.token.pk)
         self.response_302(response)
         self.assertRedirects(
             response, reverse('tokens:list'), fetch_redirect_response=False
+        )
+        self.assertEqual(
+            list(get_messages(response.wsgi_request))[0].message,
+            TOKEN_DELETE_MSG,
         )